pyproject.toml
[project]
name = "telescope"
version = "0.1.1"
description = "A framework for post-training LLMs with RL for reasoning and agents"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "datasets>=4.4.1",
    "torch>=2.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "hf-transfer>=0.1.9",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-proto>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pyarrow>=15.0.0",
    "ray>=2.9.0",
    "transformers>=4.57.3",
    "vllm>=0.12.0",
    "wandb>=0.23.1",
    "zstandard>=0.22.0",
    "liger-kernel>=0.5.10",
    "flash-attn>=2.7.0",
]

[project.optional-dependencies]
megatron = [
    "megatron-core>=0.13.0,<0.17.0",
    "megatron-bridge",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
train = "train:main"
eval = "eval:main"

[tool.uv]
# megatron-bridge==0.3.0 pulls in megatron-core[dev,mlm] whose [dev] extra
# has tightly-pinned CUDA packages (TE <2.10, opentelemetry ~=1.33, mamba-ssm,
# nvidia-modelopt, etc.) that create cascading version conflicts.
# Override megatron-core to strip the [dev] extra, and exclude CUDA-only
# packages from resolution — they are built from source in the Docker image.
# Pattern follows SkyRL and Megatron-Bridge repos.
override-dependencies = [
    "megatron-core>=0.13.0,<0.17.0",
    "transformer-engine; sys_platform == 'never'",
    "transformer-engine-torch; sys_platform == 'never'",
    "nvidia-resiliency-ext; sys_platform == 'never'",
    "mamba-ssm; sys_platform == 'never'",
    "causal-conv1d; sys_platform == 'never'",
    "flash-linear-attention; sys_platform == 'never'",
]

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telescope"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
